class_10/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.125.0",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]
